requirements.txt
crewai
crewai_tools
langchain_community
python-dotenv
streamlit
pandas
plotly
textblob
requests
holidays
pandas_market_calendars